Spine Injury Classifier Using NN Algorithm
Implement using user defined function to implement NN Algorithm classifier to classfy back injury patients in hospital based on the measurement of the shape and orientation of their pelvis and spine.
The data set contains information from **310** patients. For each patient, there are: six measurements (the x) and a label (the y). The label has **3** possible values, `’NO’` (normal), `’DH’` (herniated disk), or `’SL’` (spondilolysthesis). 
The data  is divided into a training set of 248 patients and a separate test set of 62 patients